move hookes to e2e/hooks
