+++ src/main/app/initialize.ts
@@ -5,21 +5,17 @@ import path from 'path';
 import {pathToFileURL} from 'url';
 
 import type {IpcMainInvokeEvent} from 'electron';
-import {app, BrowserWindow, ipcMain, nativeTheme, net, protocol, session, webContents} from 'electron';
+import {app, BrowserWindow, ipcMain, nativeTheme, net, protocol, session} from 'electron';
 import installExtension, {REACT_DEVELOPER_TOOLS, REDUX_DEVTOOLS} from 'electron-devtools-installer';
 import isDev from 'electron-is-dev';
 import Joi from 'joi';
 
 import MainWindow from 'app/mainWindow/mainWindow';
 import MenuManager from 'app/menus';
-import createTrayMenu from 'app/menus/tray';
 import NavigationManager from 'app/navigationManager';
-import {setUnreadBadgeSetting, setupBadge} from 'app/system/badge';
+import {setupBadge} from 'app/system/badge';
 import Tray from 'app/system/tray/tray';
-import TabManager from 'app/tabs/tabManager';
 import WebContentsManager from 'app/views/webContentsManager';
-import PopoutManager from 'app/windows/popoutManager';
-import AppState from 'common/appState';
 import {
     QUIT,
     NOTIFY_MENTION,
@@ -35,7 +31,6 @@ import {
     DOUBLE_CLICK_ON_WINDOW,
     TOGGLE_SECURE_INPUT,
     GET_APP_INFO,
-    SHOW_SETTINGS_WINDOW,
     DEVELOPER_MODE_UPDATED,
     SERVER_ADDED,
     GET_FULL_SCREEN_STATUS,
@@ -48,18 +43,16 @@ import {Logger} from 'common/log';
 import ServerManager from 'common/servers/serverManager';
 import {parseURL} from 'common/utils/url';
 import {ipcValidate} from 'common/Validator';
-import ViewManager from 'common/views/viewManager';
 import AppVersionManager from 'main/AppVersionManager';
 import AutoLauncher from 'main/AutoLauncher';
 import {configPath, updatePaths} from 'main/constants';
 import CriticalErrorHandler from 'main/CriticalErrorHandler';
 import DeveloperMode from 'main/developerMode';
-import Diagnostics from 'main/diagnostics';
 import downloadsManager from 'main/downloadsManager';
-import {registerE2eHooks} from 'main/e2e/hooks';
+import {maybeRegisterE2eHooks} from 'main/e2e/register';
 import i18nManager from 'main/i18nManager';
 import NonceManager from 'main/nonceManager';
-import notificationManager, {dispatchMentionClick, getDoNotDisturb, triggerNotificationFrameEffects} from 'main/notifications';
+import {getDoNotDisturb} from 'main/notifications';
 import parseArgs from 'main/ParseArgs';
 import PerformanceMonitor from 'main/performanceMonitor';
 import secureStorage from 'main/secureStorage';
@@ -68,7 +61,6 @@ import PermissionsManager from 'main/security/permissionsManager';
 import PreAuthManager from 'main/security/preAuthManager';
 import sentryHandler from 'main/sentryHandler';
 import SessionAttributesManager from 'main/sessionAttributes/sessionAttributesManager';
-import {installMessageBoxStub, restoreMessageBoxStub} from 'main/testMessageBoxStub';
 import updateNotifier from 'main/updateNotifier';
 import UserActivityMonitor from 'main/UserActivityMonitor';
 
@@ -80,7 +72,6 @@ import {
     handleAppWillFinishLaunching,
     handleAppWindowAllClosed,
     handleChildProcessGone,
-    certificateErrorCallbacks,
 } from './app';
 import {
     handleConfigUpdate,
@@ -99,12 +90,10 @@ import {
     handleQuit,
     handlePingDomain,
     handleToggleSecureInput,
-    handleShowSettingsModal,
 } from './intercom';
 import {
     clearAppCache,
     getDeeplinkingURL,
-    openDeepLink,
     shouldShowTrayIcon,
     updateSpellCheckerLocales,
     wasUpdated,
@@ -285,79 +274,13 @@ function initializeInterCommunicationEventListeners() {
 
     ipcMain.on(TOGGLE_SECURE_INPUT, handleToggleSecureInput);
 
-    if (process.env.NODE_ENV === 'test') {
-        ipcMain.on(SHOW_SETTINGS_WINDOW, handleShowSettingsModal);
-    }
-
     ipcMain.handle(GET_FULL_SCREEN_STATUS, (event: IpcMainInvokeEvent) => {
         return BrowserWindow.fromWebContents(event.sender)?.isFullScreen();
     });
 }
 
 async function initializeAfterAppReady() {
-    const e2eTestRefs = {
-        AppState,
-        MainWindow,
-        NotificationManager: notificationManager,
-        ServerManager,
-        TabManager,
-        ViewManager,
-        WebContentsManager,
-        Config,
-        TrayIcon: Tray,
-        Diagnostics,
-        PopoutManager,
-        updateNotifier,
-        setUnreadBadgeSetting,
-    };
-
-    registerE2eHooks({
-        e2eTestRefs,
-        openDeepLink,
-        clickTrayMenuItem: (label: string) => {
-            const truncated = label.length > 50 ? `${label.slice(0, 50)}...` : label;
-
-            function clickItem(items: Electron.MenuItem[]): boolean {
-                for (const item of items) {
-                    const itemLabel = typeof item.label === 'string' ? item.label : '';
-                    if (
-                        (itemLabel === label || itemLabel === truncated) &&
-                        item.enabled !== false &&
-                        item.visible !== false &&
-                        typeof item.click === 'function'
-                    ) {
-                        item.click();
-                        return true;
-                    }
-                    if (item.submenu?.items && clickItem(item.submenu.items)) {
-                        return true;
-                    }
-                }
-                return false;
-            }
-
-            if (!clickItem(createTrayMenu().items)) {
-                throw new Error(`Tray menu item not found: ${label}`);
-            }
-        },
-        triggerNotificationFrameEffects,
-        simulateNotificationClick: (payload) => {
-            const wc = webContents.fromId(payload.webContentsId);
-            if (!wc || wc.isDestroyed()) {
-                throw new Error(`webContents ${payload.webContentsId} is not available`);
-            }
-
-            const view = WebContentsManager.getViewByWebContentsId(wc.id);
-            if (!view) {
-                throw new Error(`No view for webContents ${payload.webContentsId}`);
-            }
-
-            dispatchMentionClick(view, wc, payload.channelId, payload.teamId, payload.url);
-        },
-        installMessageBoxStub,
-        restoreMessageBoxStub,
-        clearCertificateErrorCallbacks: () => certificateErrorCallbacks.clear(),
-    });
+    maybeRegisterE2eHooks();
 
     // Block all NTLM/Negotiate requests by default
     session.defaultSession.allowNTLMCredentialsForDomains('');
@@ -401,14 +324,6 @@ async function initializeAfterAppReady() {
     ServerManager.on(SERVER_ADDED, updateServerInfo);
     ServerManager.on(SERVER_URL_CHANGED, updateServerInfo);
     ServerManager.on(SERVER_PRE_AUTH_SECRET_CHANGED, updateServerInfo);
-
-    if (process.env.NODE_ENV === 'test') {
-        if (process.env.MM_E2E_STUB_MESSAGE_BOX === 'cancel') {
-            installMessageBoxStub([{response: 1}]);
-        } else if (process.env.MM_E2E_STUB_MESSAGE_BOX === 'trust') {
-            installMessageBoxStub([{response: 0}, {response: 0}]);
-        }
-    }
 
     ServerManager.on(SERVER_ADDED, PreAuthManager.loadPreAuthSecretForServer);
     ServerManager.init();
