+++ src/main/app/intercom.ts
@@ -7,13 +7,13 @@ import {app, BrowserWindow, Menu} from 'electron';
 import MainWindow from 'app/mainWindow/mainWindow';
 import ModalManager from 'app/mainWindow/modals/modalManager';
 import ServerViewState from 'app/serverHub';
-import {APP_MENU_WILL_CLOSE, MAIN_WINDOW_CREATED} from 'common/communication';
+import {APP_MENU_WILL_CLOSE} from 'common/communication';
 import {ModalConstants} from 'common/constants';
 import {Logger} from 'common/log';
 import ServerManager from 'common/servers/serverManager';
 import {ping} from 'common/utils/requests';
 import {parseURL} from 'common/utils/url';
-import {setTestField} from 'common/utils/util';
+import {signalE2EAppReadyWhenShown} from 'main/e2e/appReady';
 import NotificationManager from 'main/notifications';
 import {getLocalPreload} from 'main/utils';
 
@@ -89,38 +89,6 @@ export function handleMainWindowIsShown() {
     }
 
     signalE2EAppReadyWhenShown();
-}
-
-// E2E only: signals `__e2eAppReady` once the main window is visible so Playwright/Detox can wait
-// on app readiness. Gated on NODE_ENV==='test' (the same gate setTestField uses), so it adds no
-// listeners and is completely inert in normal app usage. Listener-based (no polling); also covers
-// the case where the main window has not been constructed yet.
-function signalE2EAppReadyWhenShown() {
-    if (process.env.NODE_ENV !== 'test') {
-        return;
-    }
-
-    const markReady = () => setTestField('__e2eAppReady', true);
-    const whenVisible = (win: BrowserWindow) => {
-        if (win.isVisible()) {
-            markReady();
-        } else {
-            win.once('show', markReady);
-        }
-    };
-
-    const win = MainWindow.get();
-    if (win) {
-        whenVisible(win);
-        return;
-    }
-
-    MainWindow.once(MAIN_WINDOW_CREATED, () => {
-        const created = MainWindow.get();
-        if (created) {
-            whenVisible(created);
-        }
-    });
 }
 
 export function handleWelcomeScreenModal(prefillURL?: string) {
